refactor: extra `BaseConfig` and `Extra` in dedicated `config` module
